Conflicts:
	_raw_art/editor/levels/assets/orig/pack/pack.atlas
	_raw_art/editor/levels/assets/orig/pack/pack.png
	_raw_art/editor/levels/scenes/rivertown.dt

+++ core/src/com/djammr/westernknights/entity/components/StateComponent.java
@@ -8,6 +8,6 @@ import com.djammr.westernknights.entity.EntityStates;
  */
 public class StateComponent extends Component {
 
-    public short state = EntityStates.IDLE;
+    public int state = EntityStates.IDLE;
     public boolean onGround = false;
 }
